fixing file name issues
