+++ htdocs/composer.json
@@ -18,19 +18,9 @@
     "psr-4": {
       "Oc\\": "src/Oc"
     },
-    "psr-0":{
-        "Cache\\": "src/Oc/Libse",
-        "CacheNote\\": "src/Oc/Libse",
-        "ChildWp\\": "src/Oc/Libse",
-        "Coordinate\\": "src/Oc/Libse",
-        "Language\\": "src/Oc/Libse",
-        "Validator\\": "src/Oc/Libse"
-    }
-  ,
     "classmap": [
       "lib2/",
-      "lib2/logic/",
-      "src/Oc/Libse/"
+      "lib2/logic/"
     ]
   }
 }
