+++ htdocs/src/Oc/Libse/ChildWp/EditPresenterChildWp.php
@@ -2,11 +2,6 @@
 
 namespace Oc\Libse\ChildWp;
 
-/***************************************************************************
- *  For license information see doc/license.txt
- *
- *  Unicode Reminder メモ
- ***************************************************************************/
 class EditPresenterChildWp extends PresenterChildWp
 {
     public function __construct($request = false, $translator = false)
