+++ htdocs/src/Oc/Libse/Validator/AlwaysValidValidator.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Oc\Libse\Validator;
+
+class AlwaysValidValidator
+{
+    public function isValid($value)
+    {
+        return true;
+    }
+}
