+++ htdocs/src/Oc/Libse/Validator/NumericValidator.php
@@ -2,11 +2,6 @@
 
 namespace Oc\Libse\Validator;
 
-/***************************************************************************
- *  For license information see doc/license.txt
- *
- *  Unicode Reminder メモ
- ***************************************************************************/
 abstract class NumericValidator
 {
     private $min;
@@ -27,7 +22,7 @@ abstract class NumericValidator
         }
 
         if ($this->min > $this->max) {
-            throw new InvalidArgumentException();
+            throw new \InvalidArgumentException();
         }
     }
